Move some code into src/
WIP
